# api/openapi.yaml
openapi: 3.0.3
info:
  title: identity
  version: "1"
  description: |
    Who somebody is, once, for every service operated together. Sign in with
    GitHub or Discord, get one account, carry it everywhere.

    This file is the contract. `internal/api` implements it and a test holds
    the two equal: every route is described here, every path here is a route,
    and every response the test suite provokes is validated against these
    schemas. The Go client in this package is generated from it; a TypeScript
    client would be generated from the same file.

    Two kinds of credential exist and the difference is the security model.
    An **account token** comes from signing in here (or from `POST /v1/tokens`)
    and may manage the account. An **application token** comes from
    `POST /v1/exchange` and is bound to one application's origin (its
    `token.audience`); it may call `/v1/whoami` and `/v1/avatar`, revoke
    itself, sign its browser out with `/v1/signout`, and nothing else. A
    browser-facing service must hold only its own application tokens,
    server-side, and must check `token.audience` against its own origin
    before trusting a `whoami` answer.

    An application token minted for a browser that signed in here is linked
    to that sign-in: revoking the sign-in revokes the token, and signing in to
    the same application again from the same sign-in replaces it.
servers:
  - url: /
...
security:
  - bearer: []
paths:
...
  /v1/handoff:
    post:
      operationId: handoff
      tags: [handoff]
      summary: Mint a one-time code for a consuming service's callback.
      description: |
        Called with a bearer account token, to make a credential for a service
        outside a browser. The destination must be on the redirect allowlist.
        An optional S256 PKCE challenge binds the code to the verifier the
        service holds. The token the code becomes is linked to no sign-in, so
        nothing a browser does revokes it. This service's own page mints the
        codes it hands to browsers itself.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [redirect_uri]
              properties:
                redirect_uri: { type: string, format: uri }
                code_challenge: { type: string, description: "base64url(SHA-256(verifier)), no padding." }
      responses:
        "200":
          description: A code to carry to the callback, valid for two minutes, once.
          content:
            application/json:
              schema:
                type: object
                required: [code]
                properties:
                  code: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/Failure" }
  /v1/exchange:
    post:
      operationId: exchange
      tags: [handoff]
      summary: Turn a one-time code into an application token.
      description: |
        Called server-side by the service behind the callback. The redirect
        must match the one the code was minted for, and the verifier must
        match the challenge if one was given. Any attempt consumes the code.
        A code the page minted for a signed-in browser becomes a token linked
        to that sign-in, and replaces the token that sign-in last got for the
        same application. A code minted through `POST /v1/handoff` becomes a
        token that stands alone.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, redirect_uri]
              properties:
                code: { type: string }
                redirect_uri: { type: string, format: uri }
                code_verifier: { type: string, minLength: 43, maxLength: 128 }
      responses:
        "201":
          description: A fresh token whose audience is the callback's origin.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Token" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/signout:
    post:
      operationId: signOut
      tags: [handoff]
      summary: Sign a browser out here and at every application it signed in to.
      description: |
        Called server-side by an application, with the application token of
        the session it is ending. When that token was minted for a browser
        signed in here, the sign-in is revoked, and with it every application
        token minted from it, this one included. Any other token is revoked
        alone, with anything minted from it. Answers 204 either way.
      responses:
        "204": { description: Signed out. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/Failure" }
...
  /v1/tokens/{id}:
    delete:
      operationId: revokeToken
      tags: [tokens]
      summary: Revoke one of the caller's tokens.
      description: |
        Revoking a token revokes every application token minted from it. An
        application token may revoke only itself.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Revoked. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Failure" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: "`Authorization: Bearer bsid_...`"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: "What went wrong, in a sentence." }
    AccountID:
      type: string
      description: A random 12-character hex id. Stable for the life of the account.
      pattern: "^[0-9a-f]{12}$"
      example: 3f9c2a81d04b
...
    Token:
      type: object
      description: A freshly minted credential. The token itself is shown exactly once.
      required: [token, account, handle]
      properties:
        token: { type: string, description: "`bsid_<id>_<secret>`. Empty when set as a cookie instead." }
        account: { $ref: "#/components/schemas/AccountID" }
        handle: { type: string }
        expires_at: { type: string, format: date-time }
...
  responses:
    BadRequest:
      description: The request was malformed; the error says how.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: No live bearer token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: The token exists but may not do this; the error says why.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such thing.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Failure:
      description: This service could not do it.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...